jp - opravy

+++ src/JobPanel/ui/com_manager.py
@@ -122,7 +122,8 @@ class ComManager:
         """resume first job in queue and return True else return False"""
         for  key in  self.resume_jobs:
             if not key in self._workers:
-                self.__resume_mj(key)
+                if not self.__resume_mj(key):
+                    return False
                 self._workers[key].resume()
                 return True
         return False
@@ -144,13 +145,17 @@ class ComManager:
         com_conf = comconf.CommunicatorConfig(mj_name)
         directory = inst.Installation.get_config_dir_static(mj_name, an_name)
         path = comconf.CommunicatorConfigService.get_file_path(
-            directory, comconf.CommType.app.value)        
+            directory, comconf.CommType.app.value) 
+        if not os.path.isfile(path):
+            ComWorker.get_loger().error("MultiJob settings {0} is not found".format(path))
+            return False
         with open(path, "r") as json_file:
             comconf.CommunicatorConfigService.load_file(json_file, com_conf)
             ConfigBuilder.gain_login(com_conf)
         com = Communicator(com_conf)
         worker = ComWorker(key, com)
         self._workers[key] = worker
+        return True
 
                 
     def _delete_first(self):
@@ -161,7 +166,9 @@ class ComManager:
                     self.jobs_deleted[key] = None
                     self.delete_jobs.remove(key)
                 else:
-                    self.__resume_mj(key)
+                    if not self.__resume_mj(key):
+                        self.jobs_deleted[key] = None
+                        continue
                     self._workers[key].delete()
                 return True
         return False
@@ -178,8 +185,9 @@ class ComManager:
                     self.__cancel_jobs.append(key) 
                     res = True
             else:
-                ComWorker.get_loger().error("MultiJob {0} can't be stopped, run record is not found")
-                return
+                ComWorker.get_loger().error(
+                    "MultiJob {0} can't be stopped, run record is not found".format(key))
+                return False
         return res
 
     def _terminate(self):
